+++ zospy/analyses/psf/__init__.py
@@ -1,5 +1,6 @@
 """OpticStudio analyses from the PSF category."""
 
+from zospy.analyses.psf.fft_psf import FFTPSF, FFTPSFSettings
 from zospy.analyses.psf.huygens_psf import HuygensPSF, HuygensPSFSettings
 
-__all__ = ("HuygensPSF", "HuygensPSFSettings")
+__all__ = ("FFTPSF", "FFTPSFSettings", "HuygensPSF", "HuygensPSFSettings")